+++ lib/clowne/adapters/active_record/associations/has_one.rb
@@ -5,10 +5,16 @@ module Clowne
     class ActiveRecord
       module Associations
         class HasOne < Base
+          # rubocop: disable Metrics/MethodLength
           def call(record)
             child = association
             return record unless child
-            warn '[Clowne] Has one association does not support scopes' unless scope.nil?
+            unless scope.nil?
+              warn(
+                '[Clowne] Has one association does not support scopes ' \
+                "(#{@association_name} for #{@source.class})"
+              )
+            end
 
             child_clone = clone_one(child)
             child_clone[:"#{reflection.foreign_key}"] = nil
@@ -16,6 +22,7 @@ module Clowne
 
             record
           end
+          # rubocop: enable Metrics/MethodLength
         end
       end
     end