+++ lib/clowne/adapters/active_record/associations/noop.rb
@@ -6,7 +6,10 @@ module Clowne
       module Associations
         class Noop < Base
           def call(record)
-            warn("[Clowne] Reflection #{reflection.class.name} is not supported")
+            warn(
+              "[Clowne] Reflection #{reflection.class.name} is not supported "\
+              "(#{@association_name} for #{@source.class})"
+            )
             record
           end
         end