+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change log
 
-## master branch
+## master
+
+- Support passing a block to `#clowne` for inline configuration. ([@palkan][])
+
+## 0.1.0.beta1 (2018-01-08)
 
 - Initial version. ([@ssnickolay][], [@palkan][])